chore(rust): update snapshots
Updated struct and enum fields in Rust model and SDK test snapshots to use i64 instead of i32 for integer types. Also added PartialEq, Eq, and Hash derives to several types for improved comparison and hashing. Fixed Option<String> mapping in SDK error parsing logic to avoid unnecessary allocation.

+++ generators/rust/model/src/__test__/snapshots/undiscriminated-union-types/types_error_response.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq, Hash)]
 pub struct ErrorResponse {
     pub error: String,
-    pub code: i32,
+    pub code: i64,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub details: Option<Vec<String>>,
 }
